@@ -17,8 +17,8 @@
 	"devDependencies": {
 		"@iconify/svelte": "^3.1.6",
 		"@sveltejs/adapter-auto": "^3.2.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/kit": "^2.9.0",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@types/jsonwebtoken": "^9.0.6",
 		"@types/lodash.isequal": "^4.5.8",
 		"@types/uuid": "^9.0.4",
@@ -28,15 +28,15 @@
 		"eslint": "^8.57.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.38.0",
-		"netlify-cli": "^16.3.3",
+		"netlify-cli": "^17.37.2",
 		"postcss": "^8.4.27",
 		"prettier": "^3.2.5",
 		"prettier-plugin-svelte": "^3.2.3",
-		"svelte": "^4.2.15",
+		"svelte": "^5",
 		"svelte-check": "^3.7.1",
-		"tailwindcss": "^3.4.3",
+		"tailwindcss": "^3.4.15",
 		"tslib": "^2.4.1",
-		"typescript": "^5.4.5",
+		"typescript": "^5.7.2",
 		"vite": "^5.0.0"
 	},
 	"type": "module",
@@ -47,8 +47,7 @@
 		"jsonwebtoken": "^9.0.2",
 		"lodash.isequal": "^4.5.0",
 		"mongodb": "^6.1.0",
-		"svelte-dnd-action": "^0.9.43",
-		"svelte-french-toast": "^1.2.0",
+		"svelte-dnd-action": "^0.9.60",
 		"uuid": "^9.0.0"
 	}
 }
