@@ -1,23 +1,28 @@
 <script lang="ts">
 	import Button from './Button.svelte';
 	import Dialog from './Dialog/Dialog.svelte';
-	export let onConfirm: () => void;
-	export let onOpen: () => void;
 
-	let onClose: () => void;
+	interface Props extends SvelteAllProps {
+		onOpen: () => void;
+		onConfirm: () => void;
+		isOpen: boolean;
+		onClose: () => void;
+	}
 
-	export let _onConfirm: (() => void) | undefined = () => {
+	const { onConfirm, isOpen, onClose, children }: Props = $props();
+
+	const _onConfirm: (() => void) | undefined = () => {
 		onClose();
 		onConfirm();
 	};
 </script>
 
-<Dialog bind:onOpen bind:onClose>
+<Dialog {isOpen} {onClose}>
 	<div class="flex flex-col sm:flex-row gap-3 items-center">
-		<slot />
+		{@render children()}
 		<div class="hstack gap-3">
-			<Button onClick={_onConfirm}>Yes</Button>
-			<Button onClick={onClose}>No</Button>
+			<Button onclick={_onConfirm}>Yes</Button>
+			<Button onclick={onClose}>No</Button>
 		</div>
 	</div>
 </Dialog>
