@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { LogType_enum, LogListType_enum, type MappedLog_int, type Log_int } from '$lib/types';
+	import { LogType_enum, LogListType_enum, type Log_int } from '$lib/types';
 	import {
 		clickOutside,
 		debounce,
@@ -14,8 +14,6 @@
 	import { page } from '$app/stores';
 	import { getContext, onMount, setContext } from 'svelte';
 	import { deleteLogClient, updateLogClient } from '$lib/api/logsLocalApi';
-	import toast from 'svelte-french-toast';
-	import { derived, writable, type Writable } from 'svelte/store';
 	import { isAnAutofillOpen, unfocusStoreInput, whichInputIsFocused } from '$lib/stores';
 	import { currentlyEditing, titlesAndReferences, titles } from '$lib/stores';
 	import ConfirmationDialog from '../ConfirmationDialog.svelte';
@@ -26,41 +24,98 @@
 	import LogQuestionItems from './LogQuestionItems.svelte';
 	import { useMutation, useQueryClient } from '@sveltestack/svelte-query';
 	import isEqual from 'lodash.isequal';
+	import { useDisclosure } from '$lib/hooks';
 
-	export let initialLog: Log_int;
-	const log: Writable<MappedLog_int> = writable(getMappedLog(initialLog));
-	$: initialLog, ($log = getMappedLog(initialLog));
+	interface Props {
+		initialLog: Log_int;
+		editOnMount?: boolean;
+		inputAutoFocus?: boolean;
+	}
 
-	export let editOnMount: boolean = false;
-	export let inputAutoFocus: boolean = false;
-
+	let { initialLog, editOnMount, inputAutoFocus }: Props = $props();
+	let log = $state(getMappedLog(initialLog));
+	let isAnotherCardEditing = $derived(!!$currentlyEditing && $currentlyEditing !== log.id);
+	let isEditing = $derived($currentlyEditing === log.id);
 	let container: HTMLButtonElement;
+	let focusElements: HTMLElement[] = $state([]);
 
-	const onEditLog = () => {
-		$currentlyEditing = $log.id;
-	};
+	$effect(() => {
+		// This exists because if the log is edited from outside the component, the log will not update. Example: When a log is changed and lastUpdated has been changed within a log, the logs will not order by lastUpdated as they should.
+		log = getMappedLog(initialLog);
+	});
 
-	const onStopEditing = () => {
-		$currentlyEditing = null;
-	};
-
+	const onEditLog = () => ($currentlyEditing = log.id);
+	const onStopEditing = () => ($currentlyEditing = null);
 	const onResetNewLogType: () => void = getContext('onResetNewLogType');
 	setContext('onEditLog', onEditLog);
-
-	const isEditing = derived(
-		[currentlyEditing],
-		([$currentlyEditing]) => $currentlyEditing === $log.id
-	);
 	setContext('isEditing', isEditing);
+
+	const invalidateLogs: () => Promise<void> = getContext('invalidateLogs');
+
+	const onClickInput = (e: MouseEvent) => {
+		onEditLog();
+
+		const target = e.target as HTMLElement;
+		target.focus();
+	};
+
+	const onEdit = () => {
+		if (!isAnotherCardEditing) {
+			onEditLog();
+		}
+	};
+
+	let changeReferenceInputValue: ((value: string | undefined) => void) | undefined = $state();
+
+	let onFocusAnswerInput: () => void;
+	const _onFocusAnswerInput = () => {
+		onEditLog();
+		addToEndOfRaceCondition(onFocusAnswerInput);
+	};
+
+	let haveValuesChanged = $derived(!isEqual(getMappedLog(initialLog), log));
+
+	const onTitleAutoFill = (_title: string) => {
+		const correspondingReference = $titlesAndReferences.find((t) => t.title === _title)?.reference;
+
+		changeReferenceInputValue && changeReferenceInputValue(correspondingReference ?? undefined);
+	};
+
+	const onResetChange = () => {
+		onResetNewLogType && onResetNewLogType();
+		log = getMappedLog(initialLog);
+
+		onStopEditing();
+	};
+
+	const onTextareaEnterKeydown: () => void = () => {
+		onAddItem();
+	};
+
+	const onClickOutside = () => {
+		if (isEditing) {
+			if (haveValuesChanged && !$isConfirmationDialogOpen && !$whichInputIsFocused) {
+				onOpenConfirmationDialog();
+			} else {
+				const isNewLogType = editOnMount;
+
+				if (isNewLogType) return onResetNewLogType();
+
+				onStopEditing();
+			}
+		}
+	};
 
 	const queryClient = useQueryClient();
 
 	const updateLogMutation = useMutation(updateLogClient, {
-		onMutate: async (newLog) => {
+		onMutate: async (newOrUpdatedLog) => {
 			await queryClient.cancelQueries('logs');
-			const previousLogs = queryClient.getQueryData('logs');
+			const previousLogs: Log_int[] = queryClient.getQueryData('logs') ?? [];
 			queryClient.setQueryData('logs', ({ logs: previousLogs }) => {
-				const updatedLogs = [...previousLogs, newLog];
+				const filteredLogs = previousLogs.filter((log: Log_int) => log.id !== newOrUpdatedLog.id);
+
+				const updatedLogs = [...filteredLogs, newOrUpdatedLog];
 
 				return {
 					logs: updatedLogs,
@@ -75,8 +130,7 @@
 			queryClient.setQueryData('logs', context?.previousLogs);
 		},
 		onSettled: () => {
-			queryClient.invalidateQueries('logs');
-			queryClient.invalidateQueries('allLogNotifications');
+			invalidateLogs();
 		}
 	});
 
@@ -85,7 +139,7 @@
 			await queryClient.cancelQueries('logs');
 			const previousLogs = queryClient.getQueryData('logs');
 			queryClient.setQueryData('logs', ({ logs: previousLogs }) => {
-				const updatedLogs = previousLogs.filter((log) => log.id !== logId);
+				const updatedLogs = previousLogs.filter((log: Log_int) => log.id !== logId);
 				return {
 					logs: updatedLogs,
 					count: updatedLogs.length
@@ -94,67 +148,38 @@
 
 			return { previousLogs };
 		},
-		onError: (err, newTodo, context) => {
+		onError: (err, context) => {
 			console.error('There was an error deleting log', err);
-			queryClient.setQueryData('logs', context?.previousLogs);
+			queryClient.setQueryData('logs', context?.previousLogs as Log_int[]);
 		},
 		onSettled: () => {
-			queryClient.invalidateQueries('logs');
-			queryClient.invalidateQueries('allLogNotifications');
+			invalidateLogs();
 		}
 	});
 
 	const onDelete = () => {
-		$deleteLogMutation.mutate($log.id);
+		$deleteLogMutation.mutate(log.id);
 		addToEndOfRaceCondition(onStopEditing);
 	};
 
-	const onEdit = () => {
-		const isAnotherCardEditing = !!$currentlyEditing;
-		if (!isAnotherCardEditing) {
-			onEditLog();
-		}
-	};
-
-	let changeReferenceInputValue: ((value: string | undefined) => void) | undefined = undefined;
-
-	const onTitleAutoFill = (_title: string) => {
-		const correspondingReference = $titlesAndReferences.find((t) => t.title === _title)?.reference;
-
-		changeReferenceInputValue && changeReferenceInputValue(correspondingReference ?? undefined);
-	};
-
-	let onFocusAnswerInput: () => void;
-	const _onFocusAnswerInput = () => {
-		onEditLog();
-		addToEndOfRaceCondition(onFocusAnswerInput);
-	};
-
-	const getHaveValuesChanged = () => {
-		const haveValuesChanged = !isEqual(getLogFromMappedLog($log), initialLog);
-
-		return haveValuesChanged;
-	};
-
 	const onAccept = async () => {
-		const haveValuesChanged = getHaveValuesChanged();
-
 		if (!haveValuesChanged) return onStopEditing();
 
-		if ($page.params.date && $page.params.date !== getHyphenatedStringFromDate($log.date)) {
+		if ($page.params.date && $page.params.date !== getHyphenatedStringFromDate(log.date)) {
 			const currentDate = new Date();
 			const _date = new Date(getDateFromHyphenatedString($page.params.date));
 			_date.setHours(currentDate.getHours());
 			_date.setMinutes(currentDate.getMinutes());
-			$log.date = _date;
+			log.date = _date;
 		}
 
-		const updatedLog: Log_int = getLogFromMappedLog($log);
+		log.lastUpdated = new Date();
+		const updatedLog: Log_int = getLogFromMappedLog(log);
 
 		try {
 			await $updateLogMutation.mutate(updatedLog);
 		} catch (error) {
-			toast.error('Issue updating state');
+			console.error('Issue updating state');
 		}
 
 		onResetNewLogType && onResetNewLogType();
@@ -162,84 +187,78 @@
 		addToEndOfRaceCondition(onStopEditing);
 	};
 
-	const focusElements: Writable<HTMLElement[]> = writable([]);
-
 	const onAddItem = () => {
 		onEditLog();
 
+		const doesTheLastListItemHaveValue = (() => {
+			const lastListItem = log.listItems[log.listItems.length - 1].item;
+
+			const itemValueType: Record<LogType_enum, string | number> = {
+				[LogType_enum.todo]: log.checkboxItems[log.checkboxItems.length - 1].text,
+				[LogType_enum.question]:
+					log.questions[log.questions.length - 1].question ||
+					log.questions[log.questions.length - 1].answer,
+				[LogType_enum.important]: lastListItem,
+				[LogType_enum.time]: lastListItem,
+				[LogType_enum.list]: lastListItem
+			};
+
+			return !!itemValueType[log.type];
+		})();
+
+		if (!doesTheLastListItemHaveValue) return;
+
 		const addListItem = () =>
-			$log.listItems &&
-			($log.listItems = [...$log.listItems, { id: $log.listItems.length, item: '' }]);
+			log.listItems && (log.listItems = [...log.listItems, { id: log.listItems.length, item: '' }]);
 
 		const addCheckboxItem = () =>
-			$log.checkboxItems &&
-			($log.checkboxItems = [
-				...$log.checkboxItems,
-				{ id: $log.checkboxItems.length, isChecked: false, text: '' }
+			log.checkboxItems &&
+			(log.checkboxItems = [
+				...log.checkboxItems,
+				{ id: log.checkboxItems.length, isChecked: false, text: '' }
 			]);
 
 		const addQuestionItem = () =>
-			$log.questions &&
-			($log.questions = [
-				...$log.questions,
-				{ id: $log.questions.length, question: '', answer: '' }
-			]);
+			log.questions &&
+			(log.questions = [...log.questions, { id: log.questions.length, question: '', answer: '' }]);
+
+		const isCheckboxListType = log.listType === LogListType_enum.checkbox;
 
 		const addItemTypeMethods: Record<LogType_enum, () => void> = {
-			[LogType_enum.todo]: () => {
-				addCheckboxItem();
-			},
-			[LogType_enum.question]: () => {
-				addQuestionItem();
-			},
+			[LogType_enum.todo]: addCheckboxItem,
+			[LogType_enum.question]: addQuestionItem,
 			[LogType_enum.important]: addListItem,
 			[LogType_enum.time]: addListItem,
-			[LogType_enum.list]: () => {
-				if ($log.listType === LogListType_enum.checkbox) addCheckboxItem();
-				else addListItem();
-			}
+			[LogType_enum.list]: isCheckboxListType ? addCheckboxItem : addListItem
 		};
-		addItemTypeMethods[$log.type]();
+		addItemTypeMethods[log.type]();
 
-		$focusElements = [...$focusElements, $focusElements[$focusElements.length - 1]];
+		focusElements = [...focusElements, focusElements[focusElements.length - 1]];
+
+		addToEndOfRaceCondition(() => focusElements[focusElements.length - 1].focus());
 	};
 
 	const onDeleteItem = (index: number) => {
 		const removeListItem = () =>
-			$log.listItems && ($log.listItems = $log.listItems.filter((_, i) => i !== index));
+			log.listItems && (log.listItems = log.listItems.filter((_, i) => i !== index));
 
 		const removeCheckboxItem = () =>
-			$log.checkboxItems && ($log.checkboxItems = $log.checkboxItems.filter((_, i) => i !== index));
+			log.checkboxItems && (log.checkboxItems = log.checkboxItems.filter((_, i) => i !== index));
+
 		const removeQuestionItem = () =>
-			$log.questions && ($log.questions = $log.questions.filter((_, i) => i !== index));
+			log.questions && (log.questions = log.questions.filter((_, i) => i !== index));
+
+		const isCheckboxListType = log.listType === LogListType_enum.checkbox;
 
 		const removeItemTypeMethods: Record<LogType_enum, () => void> = {
-			[LogType_enum.todo]: () => {
-				removeCheckboxItem();
-			},
-			[LogType_enum.question]: () => {
-				removeQuestionItem();
-			},
+			[LogType_enum.todo]: removeCheckboxItem,
+			[LogType_enum.question]: removeQuestionItem,
 			[LogType_enum.important]: removeListItem,
 			[LogType_enum.time]: removeListItem,
-			[LogType_enum.list]: () => {
-				if ($log.listType === LogListType_enum.checkbox) removeCheckboxItem();
-				else removeListItem();
-			}
+			[LogType_enum.list]: isCheckboxListType ? removeCheckboxItem : removeListItem
 		};
-		removeItemTypeMethods[$log.type]();
-		$focusElements = $focusElements.filter((_, i) => i !== index + 2);
-	};
-
-	const onResetChange = () => {
-		onResetNewLogType && onResetNewLogType();
-		$log = getMappedLog(initialLog);
-
-		onStopEditing();
-	};
-
-	const onTextareaEnterKeydown: () => void = () => {
-		onAddItem();
+		removeItemTypeMethods[log.type]();
+		focusElements = focusElements.filter((_, i) => i !== index + 2);
 	};
 
 	onMount(() => {
@@ -254,9 +273,7 @@
 
 		container.addEventListener('keydown', keydown);
 
-		return () => {
-			container.removeEventListener('keydown', keydown);
-		};
+		return () => container.removeEventListener('keydown', keydown);
 	});
 
 	onMount(() => {
@@ -265,19 +282,11 @@
 		}
 	});
 
-	let onOpen: () => void;
-	let isOpen: boolean;
-
-	const onClickOutside = () => {
-		if ($isEditing) {
-			const haveValuesChanged = getHaveValuesChanged();
-			if (haveValuesChanged && !isOpen && !$whichInputIsFocused) {
-				onOpen();
-			} else {
-				onStopEditing();
-			}
-		}
-	};
+	const {
+		onOpen: onOpenConfirmationDialog,
+		isOpen: isConfirmationDialogOpen,
+		onClose: onCloseConfirmationModal
+	} = useDisclosure();
 
 	const incrementDecrementPropValues: Record<LogType_enum, { min: number; max: number }> = {
 		[LogType_enum.todo]: { min: 0, max: 3 },
@@ -288,35 +297,35 @@
 	};
 
 	const onIncrement = () => {
-		if ($log.type === LogType_enum.time) {
-			$log.time = $log.time + 0.5;
+		if (log.type === LogType_enum.time) {
+			log.time = log.time + 0.5;
 
-			if (!$isEditing) {
+			if (!isEditing) {
 				debounce(() =>
 					$updateLogMutation.mutate({
-						...$log,
-						listItems: getListItemsFromMappedListItems($log.listItems)
+						...log,
+						listItems: getListItemsFromMappedListItems(log.listItems)
 					})
 				);
 			}
 		} else {
-			$log.rating = ($log.rating + 1) as 1 | 2 | 3;
+			log.rating = (log.rating + 1) as 1 | 2 | 3;
 		}
 	};
 
 	const onDecrement = () => {
-		if ($log.type === LogType_enum.time) {
-			$log.time = $log.time - 0.5;
-			if (!$isEditing) {
+		if (log.type === LogType_enum.time) {
+			log.time = log.time - 0.5;
+			if (!isEditing) {
 				debounce(() =>
 					$updateLogMutation.mutate({
-						...$log,
-						listItems: getListItemsFromMappedListItems($log.listItems)
+						...log,
+						listItems: getListItemsFromMappedListItems(log.listItems)
 					})
 				);
 			}
 		} else {
-			$log.rating = ($log.rating - 1) as 1 | 2 | 3;
+			log.rating = (log.rating - 1) as 1 | 2 | 3;
 		}
 	};
 
@@ -328,29 +337,24 @@
 		[LogType_enum.list]: ['', 'p-2 gap-1 text-sm border border-neutral-200 rounded-lg']
 	};
 
-	const onClickLog = () => {
-		if ($currentlyEditing) return;
-		onEditLog();
-	};
-
 	const onContainerKeydown = (e: KeyboardEvent) => {
-		if (!$isEditing || $isAnAutofillOpen) return;
+		if (!isEditing || $isAnAutofillOpen) return;
 
 		unfocusStoreInput();
 
-		const indexOfFocusedElement = $focusElements.indexOf(e.target as HTMLElement);
+		const indexOfFocusedElement = focusElements.indexOf(e.target as HTMLElement);
 		let indexOfNewFocusedElement: number = -1;
 
 		onKeydown(e, {
 			ArrowUp: () => {
 				if (indexOfFocusedElement === 0) {
-					indexOfNewFocusedElement = $focusElements.length - 1;
+					indexOfNewFocusedElement = focusElements.length - 1;
 				} else {
 					indexOfNewFocusedElement = indexOfFocusedElement - 1;
 				}
 			},
 			ArrowDown: () => {
-				if (indexOfFocusedElement === $focusElements.length - 1) {
+				if (indexOfFocusedElement === focusElements.length - 1) {
 					indexOfNewFocusedElement = 0;
 				} else {
 					indexOfNewFocusedElement = indexOfFocusedElement + 1;
@@ -367,97 +371,109 @@
 		});
 
 		if (indexOfNewFocusedElement > -1) {
-			$focusElements[indexOfNewFocusedElement].focus();
-			whichInputIsFocused.set($focusElements[indexOfNewFocusedElement]);
+			focusElements[indexOfNewFocusedElement].focus();
+			whichInputIsFocused.set(focusElements[indexOfNewFocusedElement]);
 		}
 	};
+
+	const incrementDecrementProps = $derived({
+		min: incrementDecrementPropValues[log.type].min,
+		max: incrementDecrementPropValues[log.type].max,
+		onIncrement,
+		onDecrement,
+		incrementDecrementValue: log.type === LogType_enum.time ? log.time : log.rating
+	});
 </script>
 
 <button
 	use:clickOutside
-	on:click_outside={onClickOutside}
-	class=""
-	on:click={onClickLog}
-	on:keydown={onContainerKeydown}
+	onclickOutside={onClickOutside}
+	onkeydown={onContainerKeydown}
 	bind:this={container}
 >
-	<div class={containerClasses[$log.type][0]}>
-		<div class="stack {containerClasses[$log.type][1]}">
-			<div class={$log.title || $log.reference || $isEditing ? 'flex' : 'hidden'}>
+	<div class={containerClasses[log.type][0]}>
+		<div class="stack {containerClasses[log.type][1]}">
+			<div class={log.title || log.reference || isEditing ? 'flex' : 'hidden'}>
 				<div class="stack gap-1 w-full">
 					<Input
-						bind:value={$log.title}
+						bind:value={log.title}
+						bind:input={focusElements[0]}
 						autofocus={inputAutoFocus}
 						placeholder="Title"
 						autofillValues={$titles}
-						isDisabled={!$isEditing}
+						isDisabled={!isEditing}
 						onAutoFill={onTitleAutoFill}
-						_class={!$isEditing && !$log.title ? 'hidden' : 'flex'}
-						bind:input={$focusElements[0]}
-						isEditing={$isEditing}
+						_class={!isEditing && !log.title ? 'hidden' : 'flex'}
+						{isEditing}
+						onclick={onClickInput}
 					/>
 					<Input
-						bind:value={$log.reference}
-						bind:changeInputValue={changeReferenceInputValue}
+						bind:value={log.reference}
+						bind:input={focusElements[1]}
 						placeholder="Reference"
-						isDisabled={!$isEditing}
-						_class={!$isEditing && !$log.reference ? 'hidden' : 'flex'}
-						bind:input={$focusElements[1]}
-						isEditing={$isEditing}
+						isDisabled={!isEditing}
+						_class={!isEditing && !log.reference ? 'hidden' : 'flex'}
+						{isEditing}
+						onclick={onClickInput}
 					/>
 				</div>
 			</div>
 			<div class="hstack center gap-2">
-				{#if $log.type === LogType_enum.todo || ($log.type === LogType_enum.list && $log.listType === LogListType_enum.checkbox)}
+				<!-- todo: All 3 of these conditionally rendered items should reduce to one once LogQuestionItems and LogCheckboxItems has been refactored into LogListItems -->
+				{#if log.type === LogType_enum.todo || (log.type === LogType_enum.list && log.listType === LogListType_enum.checkbox)}
 					<div class="flex-1">
 						<CheckboxItems
-							bind:checkboxes={$log.checkboxItems}
+							bind:checkboxes={log.checkboxItems}
+							bind:focusElements
 							{isEditing}
 							onEnterKeydown={onTextareaEnterKeydown}
 							onDeleteBullet={onDeleteItem}
 							{onEdit}
-							{focusElements}
+							{onClickInput}
 						/>
 					</div>
-				{:else if $log.type === LogType_enum.question}
+				{:else if log.type === LogType_enum.question}
 					<LogQuestionItems
-						bind:questions={$log.questions}
+						bind:questions={log.questions}
+						bind:focusElements
 						onFocusAnswerInput={_onFocusAnswerInput}
 						{isEditing}
 						onDeleteQuestion={onDeleteItem}
-						id={$log.id}
-						{focusElements}
+						id={log.id}
+						{onClickInput}
 					/>
-				{:else if $log.type === LogType_enum.important || $log.type === LogType_enum.time || ($log.type === LogType_enum.list && $log.listType !== LogListType_enum.checkbox)}
+				{:else if log.type === LogType_enum.important || log.type === LogType_enum.time || (log.type === LogType_enum.list && log.listType !== LogListType_enum.checkbox)}
 					<ListItems
-						bind:items={$log.listItems}
-						bind:listType={$log.listType}
+						bind:items={log.listItems}
+						bind:focusElements
+						listType={log.listType}
 						{isEditing}
 						onEnterKeydown={onTextareaEnterKeydown}
 						{onDeleteItem}
-						logType={$log.type}
-						{focusElements}
+						logType={log.type}
+						{onClickInput}
 					/>
 				{/if}
 			</div>
 			<BottomOptions
-				incrementDecrementProps={{
-					min: incrementDecrementPropValues[$log.type].min,
-					max: incrementDecrementPropValues[$log.type].max,
-					onIncrement,
-					onDecrement
-				}}
-				incrementDecrementValue={$log.type === LogType_enum.time ? $log.time : $log.rating}
-				isEditing={$isEditing}
+				bind:log
+				{incrementDecrementProps}
+				{isEditing}
 				{onAccept}
 				{onAddItem}
 				{onDelete}
-				{log}
+				{onEditLog}
+				{onStopEditing}
 			/>
 		</div>
 	</div>
 </button>
 
-<ConfirmationDialog onConfirm={onResetChange} bind:onOpen
-	>Did you want to reset your changes?</ConfirmationDialog
+<ConfirmationDialog
+	onConfirm={onResetChange}
+	onOpen={onOpenConfirmationDialog}
+	isOpen={$isConfirmationDialogOpen}
+	onClose={onCloseConfirmationModal}
+>
+	Did you want to reset your changes?</ConfirmationDialog
 >
