@@ -6,7 +6,6 @@
 	import Log from '$lib/components/Logs/Log.svelte';
 	import Search from '$lib/components/Search.svelte';
 	import { icons } from '$lib/general/icons';
-	import { viewport } from '$lib/hooks';
 	import { searchValue, whichInputIsFocused } from '$lib/stores';
 	import { allLogs, searchableInputs } from '$lib/types';
 	import { arraysAreEqual } from '$lib/utils/arrays';
@@ -19,50 +18,58 @@
 	import Icon from '@iconify/svelte';
 	import { useInfiniteQuery, useQueryClient } from '@sveltestack/svelte-query';
 	import { onMount } from 'svelte';
-	import { derived, writable, type Writable } from 'svelte/store';
 	import { axios } from '$lib/general';
+	import { useDisclosure, viewport } from '$lib';
 
 	const queryClient = useQueryClient();
+	const _searchParams = $derived(new URLSearchParams($page.url.searchParams));
 
-	const searchParams = derived(page, ($page) => {
-		const searchParams = new URLSearchParams($page.url.searchParams);
-
-		const string = searchParams.toString();
-
-		const array = searchParamsStringToEntriesArray(searchParams);
-
-		return { string, array };
+	let searchParams = $derived({
+		string: _searchParams.toString(),
+		array: searchParamsStringToEntriesArray(_searchParams)
 	});
 
-	const filters: Writable<Array<Array<string>>> = writable($searchParams.array);
+	let filters = $derived(searchParams.array);
 
-	const filtersValues = derived(filters, ($filters) => {
-		const string = entriesArrayToSearchParamsString($filters as Array<[string, string]>);
-		return { string };
-	});
+	const onClearFilters = () => goto(`/${$page.params.space}/filter`, { keepFocus: true });
+
+	const onAddFilter = (filter: [string, string]) =>
+		goto(
+			`/${$page.params.space}/filter?${entriesArrayToSearchParamsString([...filters, filter])}`,
+			{ keepFocus: true }
+		);
+
+	const onRemoveFilter = (filter: [string, string]) =>
+		goto(
+			`/${$page.params.space}/filter?${entriesArrayToSearchParamsString(filters.filter((f) => !arraysAreEqual(f, filter)))}`,
+			{ keepFocus: true }
+		);
+
+	const setFilters = (filters: [string, string][]) =>
+		goto(`/${$page.params.space}/filter?${entriesArrayToSearchParamsString(filters)}`, {
+			keepFocus: true
+		});
 
 	const onClickClear = () => {
-		$filters = [];
+		onClearFilters();
 		queryClient.invalidateQueries('filteredLogs');
 		$searchValue = '';
 	};
 
-	let hasPageLoaded = false;
-	let onSearchInputFocus: () => void;
+	let hasPageLoaded = $state(false);
 
 	onMount(() => {
 		hasPageLoaded = true;
 		const hasSearchQuery = $page.url.searchParams.has('search');
 
 		if (hasSearchQuery) {
-			onSearchInputFocus();
 			$searchValue = $page.url.searchParams.get('search') ?? '';
 		}
 	});
 
-	let timer;
+	let timer: any = $state(undefined);
 
-	const debounce = (fn: () => any, delay = 500) => {
+	const debounce = (fn: () => void, delay = 500) => {
 		const timeout = () => {
 			clearTimeout(timer);
 			timer = setTimeout(() => {
@@ -72,22 +79,32 @@
 		timeout();
 	};
 
-	$: $filters,
-		hasPageLoaded &&
-			debounce(() => {
-				const url = $filtersValues.string
-					? `${$page.url.pathname}?${$filtersValues.string}`
-					: `${$page.url.pathname}`;
+	// I dont know what this is for. It may have something to do with the filter page being loaded with a search query. I think come back to this once you have a better understanding of the codebase.
 
-				goto(url, { keepFocus: true });
-			});
+	// let filtersValues = $derived({
+	// 	string: (() => {
+	// 		const string = entriesArrayToSearchParamsString(filters as Array<[string, string]>);
+	// 		return { string };
+	// 	})()
+	// });
+
+	// $effect(() => {
+	// hasPageLoaded &&
+	// 	filters &&
+	// 	debounce(() => {
+	// 		const url = filtersValues.string
+	// 			? `${$page.url.pathname}?${filtersValues.string}`
+	// 			: `${$page.url.pathname}`;
+	// 		goto(url, { keepFocus: true });
+	// 	});
+	// });
 
 	const fetchFilteredLogs = async ({ pageParam: skip = 0 }) => {
 		const result = await axios.get(`/log`, {
 			params: {
 				space: replaceAllSpacesWithHyphens($page.params.space),
 				skip,
-				json: JSON.stringify($searchParams.array)
+				json: JSON.stringify(searchParams.array)
 			}
 		});
 
@@ -104,26 +121,28 @@
 		}
 	});
 
-	const updatedStateAndInvalidate = () => {
-		$filters = searchParamsStringToEntriesArray($page.url.searchParams);
-		queryClient.invalidateQueries('filteredLogs');
-	};
-	$: $page.url, updatedStateAndInvalidate();
+	$effect(() => {
+		const updatedStateAndInvalidate = () => {
+			setFilters(searchParams.array);
+			queryClient.invalidateQueries('filteredLogs');
+		};
 
-	let headerContainerHeight: number;
-	let parentContainerHeight: number;
-	$: logContainerHeight = parentContainerHeight - headerContainerHeight - 20;
+		$page.url && updatedStateAndInvalidate();
+	});
+
+	let headerContainerHeight: number = $state(0);
+	let parentContainerHeight: number = $state(0);
+
+	const logContainerHeight: number = $derived(parentContainerHeight - headerContainerHeight - 20);
 
 	onMount(() => {
-		const onKeydown = (e) => {
-			if (e.key === 'Escape') {
-				goto(`/${$page.params.space}`);
-			}
+		const onKeydown = (e: KeyboardEvent) => {
+			if (e.key === 'Escape') goto(`/${$page.params.space}`, { keepFocus: true });
 		};
+
 		document.addEventListener('keydown', onKeydown);
-		return () => {
-			document.removeEventListener('keydown', onKeydown);
-		};
+
+		return () => document.removeEventListener('keydown', onKeydown);
 	});
 
 	const onGetNextPage = () => {
@@ -133,49 +152,58 @@
 	};
 
 	const onFilterChange = (filterValue: string[]) => {
-		const hasExistingValue = $filters.some((filter) => arraysAreEqual(filter, filterValue));
+		const hasMatchingFilter = filters.some((filter) => arraysAreEqual(filter, filterValue));
 
-		if (hasExistingValue) {
-			$filters = $filters.filter((filter) => !arraysAreEqual(filter, filterValue));
+		if (hasMatchingFilter) {
+			onRemoveFilter(filterValue as [string, string]);
 		} else {
-			$filters = [...$filters, filterValue];
+			onAddFilter(filterValue as [string, string]);
 		}
 	};
 
 	const onSearch = () => {
-		if ($filters.some((filter) => filter[0] === 'search')) {
-			$filters = $filters.filter((filter) => filter[0] !== 'search');
-		}
+		const removeAllSearchFilters = () =>
+			goto(
+				`/${$page.params.space}/filter?${entriesArrayToSearchParamsString(filters.filter((filter) => filter[0] !== 'search'))}`,
+				{ keepFocus: true }
+			);
 
-		if ($searchValue) {
-			$filters = [...$filters, ['search', $searchValue]];
-		} else {
-			$filters = $filters.filter((filter) => filter[0] !== 'search');
-		}
+		const isThereAnExistingSearchFilter = filters.some((filter) => filter[0] === 'search');
+
+		if (isThereAnExistingSearchFilter) removeAllSearchFilters();
+
+		if ($searchValue) onAddFilter(['search', $searchValue]);
+		else removeAllSearchFilters();
 	};
-	let onCloseExport: () => void;
-	let onOpenExport: () => void;
+
+	const {
+		isOpen: isExportDialogOpen,
+		onOpen: openExportDialog,
+		onClose: closeExportDialog
+	} = useDisclosure();
+
+	const onSearchChange = (e: Event) => {
+		const target = e.target as HTMLInputElement;
+		const newValue = target.value;
+		$searchValue = newValue;
+
+		debounce(() => onSearch());
+	};
 </script>
 
-<div class="stack flex-1 gap-3" bind:clientHeight={parentContainerHeight}>
-	<div class="center stack gap-2" bind:clientHeight={headerContainerHeight}>
+<div bind:clientHeight={parentContainerHeight} class="stack flex-1 gap-3">
+	<div bind:clientHeight={headerContainerHeight} class="center stack gap-2">
 		<div class="hstack gap-3 flex-wrap center">
 			<div class="hstack gap-3 border border-gray-100 py-1 px-2 rounded-md flex-wrap center">
-				<Search
-					bind:value={$searchValue}
-					onChange={(e) => debounce(() => onSearch())}
-					{onClickClear}
-					showEnter={false}
-					bind:onFocus={onSearchInputFocus}
-				/>
+				<Search value={$searchValue} onchange={onSearchChange} {onClickClear} showEnter={false} />
 				<div class="hstack gap-2 {$searchValue ? 'opacity-100' : 'opacity-20'}">
 					{#each searchableInputs as key}
 						<label class="capitalize text-xs center gap-1 hstack gap-[2px]">
 							{camelCaseToLower(key)}
 							<input
 								type="checkbox"
-								on:change={() => onFilterChange(['searchType', key])}
-								checked={$filters.some((filterArray) =>
+								onchange={() => onFilterChange(['searchType', key])}
+								checked={filters.some((filterArray) =>
 									arraysAreEqual(filterArray, ['searchType', key])
 								)}
 								disabled={!$searchValue}
@@ -183,7 +211,7 @@
 						</label>
 					{/each}
 				</div>
-				<Button onClick={onOpenExport}>
+				<Button onclick={openExportDialog}>
 					<Icon icon={icons.export} height="20px" class="text-gray-400" />
 				</Button>
 			</div>
@@ -193,8 +221,8 @@
 						{camelCaseToLower(key)}
 						<input
 							type="checkbox"
-							on:change={() => onFilterChange(['type', key])}
-							checked={$filters.some((filterArray) => arraysAreEqual(filterArray, ['type', key]))}
+							onchange={() => onFilterChange(['type', key])}
+							checked={filters.some((filterArray) => arraysAreEqual(filterArray, ['type', key]))}
 						/>
 					</label>
 				{/each}
@@ -210,7 +238,7 @@
 		>
 			{#if $filteredLogsQuery.isLoading}
 				{#each Array(5) as _}
-					<div class="bg-neutral-100 rounded-sm h-[120px] w-full" />
+					<div class="bg-neutral-100 rounded-sm h-[120px] w-full"></div>
 				{/each}
 			{:else if $filteredLogsQuery.isError}
 				Error
@@ -224,7 +252,7 @@
 					{#if $filteredLogsQuery.isFetching}
 						Loading more...
 					{:else if $filteredLogsQuery.hasNextPage}
-						<div use:viewport on:enterViewport={onGetNextPage} class="h-10 w-full" />
+						<div use:viewport onenterViewport={onGetNextPage} class="h-10 w-full" />
 					{:else}
 						{$filteredLogsQuery.data.pages[0].data.total} Results
 					{/if}
@@ -235,8 +263,9 @@
 </div>
 
 <ExportDialog
-	bind:onClose={onCloseExport}
-	bind:onOpen={onOpenExport}
+	isOpen={$isExportDialogOpen}
+	onOpen={openExportDialog}
+	onClose={closeExportDialog}
 	logsData={$filteredLogsQuery.data}
 	isLoadingLogs={$filteredLogsQuery.isLoading}
 	isLogsError={$filteredLogsQuery.isError}
