@@ -3,8 +3,9 @@ import {
 	getAllLogNotifications,
 	getAndCheckCollectionFromToken
 } from '$lib/server/index.js';
+import type { RequestEvent } from '@sveltejs/kit';
 
-export const GET = async ({ request, url: { searchParams } }) =>
+export const GET = async ({ request, url: { searchParams } }: RequestEvent) =>
 	checkAccessTokenMiddleware(request, async () =>
 		getAndCheckCollectionFromToken(request, async (collection) => {
 			const result = new URLSearchParams(searchParams).entries();
